<docs>: aprill week1 problem solve

+++ delete_exe.py
@@ -5,6 +5,8 @@ isEmpty = True
 for file in files:
     if '.' not in file:
         isEmpty = False
+        if os.path.isdir(file):
+            continue
         print(file, 'is deleted')
         os.remove(file)
 if isEmpty:
